Unify reward and loss plots
