+++ js/platform/base.js
@@ -1,4 +1,5 @@
 import ScatterRenderer from '../renderer/scatter.js'
+import LinePlotter from '../renderer/util/lineplot.js'
 
 import { DataPointStarPlotter, specialCategory, DataPoint, DataLine } from '../utils.js'
 import TableRenderer from '../renderer/table.js'
@@ -267,7 +268,7 @@ export class DefaultPlatform extends BasePlatform {
 		if (!this._loss) {
 			const orgText = this.setting.footer.innerText
 			this.setting.footer.innerText = ''
-			this._loss = new LossPlotter(this, this.setting.footer)
+			this._loss = new LinePlotter(this.setting.footer)
 			this._getEvaluateElm().innerText = orgText
 		}
 		this._loss.add(value)
@@ -278,198 +279,5 @@ export class DefaultPlatform extends BasePlatform {
 		this.setting.footer.innerText = ''
 		super.terminate()
 		this._tablerenderer.terminate()
-	}
-}
-
-export class LossPlotter {
-	constructor(platform, r) {
-		this._platform = platform
-		this._r = r
-
-		this._item = null
-	}
-
-	add(value) {
-		if (!this._item) {
-			if (typeof value === 'object') {
-				this._item = {}
-				for (const key of Object.keys(value)) {
-					this._item[key] = new LossPlotterItem(this._platform, this._r)
-					this._item[key].name = key
-				}
-			} else {
-				this._item = new LossPlotterItem(this._platform, this._r)
-			}
-		}
-		if (typeof value === 'object') {
-			for (const key of Object.keys(value)) {
-				this._item[key].add(value[key])
-			}
-		} else {
-			this._item.add(value)
-		}
-	}
-
-	terminate() {
-		if (this._item instanceof LossPlotterItem) {
-			this._item.terminate()
-		} else {
-			for (const key of Object.keys(this._item)) {
-				this._item[key].terminate()
-			}
-		}
-	}
-}
-
-class LossPlotterItem {
-	constructor(platform, r) {
-		this._platform = platform
-		this._root = document.createElement('span')
-		r.appendChild(this._root)
-		this._caption = document.createElement('div')
-		this._caption.innerText = 'loss'
-		this._root.appendChild(this._caption)
-
-		const cont = document.createElement('span')
-		cont.style.display = 'inline-flex'
-		cont.style.alignItems = 'flex-start'
-		this._root.appendChild(cont)
-
-		const svg = document.createElementNS('http://www.w3.org/2000/svg', 'svg')
-		svg.setAttribute('width', 200)
-		svg.setAttribute('height', 50)
-		cont.appendChild(svg)
-
-		cont.appendChild(document.createTextNode('scale:'))
-		this._scale = document.createElement('select')
-		for (const name of ['linear', 'log']) {
-			const opt = document.createElement('option')
-			opt.value = name
-			opt.innerText = name
-			this._scale.appendChild(opt)
-		}
-		this._scale.onchange = () => this.plotRewards()
-		cont.appendChild(this._scale)
-
-		this._stats = document.createElement('span')
-		this._stats.style.display = 'inline-flex'
-		this._stats.style.flexDirection = 'column'
-		this._stats.style.fontSize = '80%'
-		cont.appendChild(this._stats)
-		for (const k of ['max', 'ave', 'min']) {
-			const txt = document.createElement('span')
-			txt.classList.add(k + 'txt')
-			this._stats.append(txt)
-		}
-
-		this._plot_count = 10000
-		this._print_count = 10
-		this._plot_smooth_window = 20
-
-		this._history = []
-	}
-
-	set name(value) {
-		this._caption.innerText = value
-	}
-
-	add(value) {
-		this._history.push(value)
-		this.plotRewards()
-	}
-
-	terminate() {
-		this._root.remove()
-	}
-
-	lastHistory(length = 0) {
-		if (length <= 0) {
-			return this._history
-		}
-		const historyLength = this._history.length
-		return this._history.slice(Math.max(0, historyLength - length), historyLength)
-	}
-
-	plotRewards() {
-		const svg = this._root.querySelector('svg')
-		const width = svg.width.baseVal.value
-		const height = svg.height.baseVal.value
-		let path = null
-		let sm_path = null
-		const mintxt = this._stats.querySelector('.mintxt')
-		const maxtxt = this._stats.querySelector('.maxtxt')
-		const avetxt = this._stats.querySelector('.avetxt')
-		if (svg.childNodes.length === 0) {
-			path = document.createElementNS('http://www.w3.org/2000/svg', 'path')
-			path.setAttribute('name', 'value')
-			path.setAttribute('stroke', 'black')
-			path.setAttribute('fill-opacity', 0)
-			svg.appendChild(path)
-			sm_path = document.createElementNS('http://www.w3.org/2000/svg', 'path')
-			sm_path.setAttribute('name', 'smooth')
-			sm_path.setAttribute('stroke', 'green')
-			sm_path.setAttribute('fill-opacity', 0)
-			svg.appendChild(sm_path)
-		} else {
-			path = svg.querySelector('path[name=value]')
-			sm_path = svg.querySelector('path[name=smooth]')
-		}
-
-		const lastHistory = this.lastHistory(this._plot_count)
-		if (lastHistory.length === 0) {
-			svg.style.display = 'none'
-			path.removeAttribute('d')
-			sm_path.removeAttribute('d')
-			return
-		} else {
-			svg.style.display = null
-		}
-		const maxr = Math.max(...lastHistory)
-		const minr = Math.min(...lastHistory)
-
-		const fmtNum = f => {
-			if (typeof f !== 'number') {
-				return f
-			}
-			const scale = -Math.floor(Math.log10(Math.abs(f))) + 3
-			return Math.round(f * 10 ** scale) / 10 ** scale
-		}
-
-		mintxt.innerText = `Min: ${fmtNum(minr)}`
-		maxtxt.innerText = `Max: ${fmtNum(maxr)}`
-		if (maxr === minr) return
-
-		const pp = (i, v) => {
-			if (this._scale.value === 'log') {
-				return [
-					(width * i) / (lastHistory.length - 1),
-					(1 - (Math.log(v) - Math.log(minr)) / (Math.log(maxr) - Math.log(minr))) * height,
-				]
-			}
-			return [(width * i) / (lastHistory.length - 1), (1 - (v - minr) / (maxr - minr)) * height]
-		}
-
-		const p = lastHistory.map((v, i) => pp(i, v))
-		const line = p => {
-			let s = ''
-			for (let i = 0; i < p.length; i++) {
-				s += `${i === 0 ? 'M' : 'L'}${p[i][0]},${p[i][1]}`
-			}
-			return s
-		}
-		path.setAttribute('d', line(p))
-
-		const smp = []
-		for (let i = 0; i < lastHistory.length - this._plot_smooth_window; i++) {
-			let s = 0
-			for (let k = 0; k < this._plot_smooth_window; k++) {
-				s += lastHistory[i + k]
-			}
-			smp.push([i + this._plot_smooth_window, s / this._plot_smooth_window])
-		}
-		if (smp.length > 0) {
-			sm_path.setAttribute('d', line(smp.map(p => pp(...p))))
-			avetxt.innerText = `Mean(${this._plot_smooth_window}): ${fmtNum(smp[smp.length - 1]?.[1])}`
-		}
 	}
 }
